// src/app/core/models/local-event.model.ts
export interface LocalEvent {
  id: string;

  name: string;

  url?: string;

  start?: string;

  status?: string;

  venue?: string;

  city?: string;

  address?: string;

  segment?: string;

  genre?: string;

  subGenre?: string;

  image?: string;

  priceMin?: number;

  priceMax?: number;

  currency?: string;
}